+++ packages/frontend/app/dashboard/components/information-card.tsx
