+++ packages/frontend/app/dashboard/components/public-witnessing-schedule-card.tsx
